requirements.txt
aiohttp>=3.8.1
PyJWT>=2.3.0
